backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiogram==3.7.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis[hiredis]==5.0.4
celery==5.4.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
Pillow==10.3.0
